agents/a2a-sidecar/requirements-dev.txt
# Runtime dependency
-r requirements.txt

# Test dependencies
pytest>=8.0
pytest-asyncio>=0.23
pytest-aiohttp>=1.0
